DataProcessing/requirements.txt
azure-common==1.1.23
azure-core==1.0.0b4
azure-eventhub==5.0.0b4
certifi==2019.9.11
chardet==3.0.4
idna==2.8
numpy==1.17.3
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
uamqp==1.2.3
urllib3==1.25.6
